Add decimal, min, max and step to widget money
